+++ bindings/node/lib/clientEncryption.js
@@ -13,7 +13,7 @@ module.exports = function (modules) {
     MongoCryptCreateEncryptedCollectionError,
     MongoCryptCreateDataKeyError
   } = require('./errors');
-  const { loadCredentials } = require('./credentialsProvider');
+  const { loadCredentials } = require('./providers');
   const cryptoCallbacks = require('./cryptoCallbacks');
   const { promisify } = require('util');
