+++ bindings/node/test/providers/credentialsProvider.test.js
@@ -2,8 +2,8 @@
 
 const { expect } = require('chai');
 const http = require('http');
-const requirements = require('./requirements.helper');
-const { loadCredentials, isEmptyCredentials } = require('../lib/credentialsProvider');
+const requirements = require('../requirements.helper');
+const { loadCredentials, isEmptyCredentials } = require('../../lib/providers');
 
 const originalAccessKeyId = process.env.AWS_ACCESS_KEY_ID;
 const originalSecretAccessKey = process.env.AWS_SECRET_ACCESS_KEY;
