+++ bindings/node/test/release.test.js
@@ -16,7 +16,9 @@ const REQUIRED_FILES = [
   'package/lib/buffer_pool.js',
   'package/lib/clientEncryption.js',
   'package/lib/common.js',
-  'package/lib/credentialsProvider.js',
+  'package/lib/providers/index.js',
+  'package/lib/providers/gcp.js',
+  'package/lib/providers/aws.js',
   'package/lib/cryptoCallbacks.js',
   'package/lib/errors.js',
   'package/lib/mongocryptdManager.js',
